Revert "x86/HPET: don't enable legacy replacement mode unconditionally"
This reverts commit e680cc48b7184d3489873d6776f84ba1fc238ced.

It was committed despite multiple objections.  The agreed upon fix is a
different variation of the same original patch, and the delta between the two
is far from clear.

By reverting this commit first, the fixes are clear and coherent as individual
patches, and in the appropriate form for backport to the older trees.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>
master commit: 2b05d427a09e108589dc57f54b517d0ffdf973ef
master date: 2021-04-15 18:25:40 +0100